Add a new DescribesDataColumns
