+++ base/src/main/java/uk/ac/ebi/atlas/model/DescribesDataColumns.java
@@ -1,5 +1,6 @@
 package uk.ac.ebi.atlas.model;
 
+import java.util.Objects;
 import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
@@ -18,4 +19,15 @@ public abstract class DescribesDataColumns {
     }
 
     public abstract Set<String> assaysAnalyzedForThisDataColumn();
+
+    @Override
+    public int hashCode() {return id.hashCode();}
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {return true;}
+        if (obj == null || getClass() != obj.getClass()) {return false;}
+        final DescribesDataColumns other = (DescribesDataColumns) obj;
+        return Objects.equals(this.id, other.id);
+    }
 }
